chore: applying some fixes noticed whilst implementing ts transact (#245)
* chore: applying some fixes noticed whilst implementing ts transact

+++ crates/algod_client/src/models/application_state_schema.rs
@@ -17,10 +17,10 @@ use serde::{Deserialize, Serialize};
 pub struct ApplicationStateSchema {
     /// \[nui\] num of uints.
     #[serde(rename = "num-uint")]
-    pub num_uint: u64,
+    pub num_uint: u32,
     /// \[nbs\] num of byte slices.
     #[serde(rename = "num-byte-slice")]
-    pub num_byte_slice: u64,
+    pub num_byte_slice: u32,
 }
 
 impl AlgorandMsgpack for ApplicationStateSchema {
@@ -29,7 +29,7 @@ impl AlgorandMsgpack for ApplicationStateSchema {
 
 impl ApplicationStateSchema {
     /// Constructor for ApplicationStateSchema
-    pub fn new(num_uint: u64, num_byte_slice: u64) -> ApplicationStateSchema {
+    pub fn new(num_uint: u32, num_byte_slice: u32) -> ApplicationStateSchema {
         ApplicationStateSchema {
             num_uint,
             num_byte_slice,
