+++ crates/algokit_transact/src/constants.rs
@@ -13,7 +13,7 @@ pub const EMPTY_SIGNATURE: [u8; ALGORAND_SIGNATURE_BYTE_LENGTH] =
     [0; ALGORAND_SIGNATURE_BYTE_LENGTH];
 
 // Application program size constraints
-pub const MAX_EXTRA_PROGRAM_PAGES: u64 = 3;
+pub const MAX_EXTRA_PROGRAM_PAGES: u32 = 3;
 pub const PROGRAM_PAGE_SIZE: usize = 2048; // In bytes
 
 // Application reference limits
@@ -26,5 +26,5 @@ pub const MAX_ASSET_REFERENCES: usize = 8;
 pub const MAX_BOX_REFERENCES: usize = 8;
 
 // Application state schema limits
-pub const MAX_GLOBAL_STATE_KEYS: u64 = 64;
-pub const MAX_LOCAL_STATE_KEYS: u64 = 16;
+pub const MAX_GLOBAL_STATE_KEYS: u32 = 64;
+pub const MAX_LOCAL_STATE_KEYS: u32 = 16;
